compact storing for blocknumber and oracle data

+++ contracts/interfaces/IUniswapV2.sol
@@ -35,7 +35,7 @@ interface IUniswapV2 {
 
     function getReserves() external view returns (uint128, uint128);
     function getAccumulatedPrices() external view returns (uint256, uint256);
-    function getBlockNumberLast() external view returns (uint256);
+    function getBlockNumberLast() external view returns (uint32);
 
     function getAmountOutput(uint128 amountInput, uint128 reserveInput, uint128 reserveOutput)
         external pure returns (uint128 amountOutput);
